Remove flux stores
