+++ components/create_post/index.js
@@ -28,7 +28,7 @@ import {Posts, Preferences as PreferencesRedux} from 'mattermost-redux/constants
 
 import {connectionErrorCount} from 'selectors/views/system';
 
-import {emitUserPostedEvent, postListScrollChangeToBottom} from 'actions/global_actions.jsx';
+import {postListScrollChangeToBottom} from 'actions/global_actions.jsx';
 import {addReaction, createPost, setEditingPost} from 'actions/post_actions.jsx';
 import {selectPostFromRightHandSideSearchByPostId} from 'actions/views/rhs';
 import {executeCommand} from 'actions/command';
@@ -95,7 +95,6 @@ function mapStateToProps() {
 
 function onSubmitPost(post, fileInfos) {
     return (dispatch) => {
-        emitUserPostedEvent(post);
         dispatch(createPost(post, fileInfos));
         postListScrollChangeToBottom();
     };
