+++ utils/constants.jsx
@@ -93,34 +93,8 @@ export const Preferences = {
 };
 
 export const ActionTypes = keyMirror({
-    RECEIVED_ERROR: null,
-    CLICK_CHANNEL: null,
-    CREATE_CHANNEL: null,
-    CREATE_POST: null,
-    CREATE_COMMENT: null,
-    POST_DELETED: null,
-    POST_UPDATED: null,
-    REMOVE_POST: null,
-
-    RECEIVED_CHANNELS: null,
-    RECEIVED_CHANNEL: null,
-    RECEIVED_CHANNEL_MEMBER: null,
-    RECEIVED_MORE_CHANNELS: null,
-    RECEIVED_CHANNEL_STATS: null,
-    RECEIVED_MY_CHANNEL_MEMBERS: null,
-    RECEIVED_MEMBERS_IN_CHANNEL: null,
-
-    FOCUS_POST: null,
-    RECEIVED_POSTS: null,
     RECEIVED_FOCUSED_POST: null,
-    RECEIVED_EDIT_POST: null,
-    EDIT_POST: null,
     SELECT_POST: null,
-    RECEIVED_POST_SELECTED: null,
-    RECEIVED_MENTION_DATA: null,
-    RECEIVED_ADD_MENTION: null,
-    RECEIVED_POST_PINNED: null,
-    RECEIVED_POST_UNPINNED: null,
     INCREASE_POST_VISIBILITY: null,
     LOADING_POSTS: null,
 
@@ -141,75 +115,10 @@ export const ActionTypes = keyMirror({
     SEARCH_PINNED_POSTS_SUCCESS: null,
     SEARCH_PINNED_POSTS_FAILURE: null,
 
-    RECEIVED_PROFILES: null,
-    RECEIVED_PROFILES_IN_TEAM: null,
-    RECEIVED_PROFILES_NOT_IN_TEAM: null,
-    RECEIVED_PROFILE: null,
-    RECEIVED_PROFILES_IN_CHANNEL: null,
-    RECEIVED_PROFILES_NOT_IN_CHANNEL: null,
-    RECEIVED_PROFILES_WITHOUT_TEAM: null,
-    RECEIVED_ME: null,
-    RECEIVED_SESSIONS: null,
-    RECEIVED_AUDITS: null,
-    RECEIVED_TEAMS: null,
-    RECEIVED_STATUSES: null,
-    RECEIVED_PREFERENCE: null,
-    RECEIVED_PREFERENCES: null,
-    DELETED_PREFERENCES: null,
-    RECEIVED_FILE_INFOS: null,
-    RECEIVED_ANALYTICS: null,
-
-    RECEIVED_INCOMING_WEBHOOKS: null,
-    RECEIVED_INCOMING_WEBHOOK: null,
-    UPDATED_INCOMING_WEBHOOK: null,
-    REMOVED_INCOMING_WEBHOOK: null,
-    RECEIVED_OUTGOING_WEBHOOKS: null,
-    RECEIVED_OUTGOING_WEBHOOK: null,
-    UPDATED_OUTGOING_WEBHOOK: null,
-    REMOVED_OUTGOING_WEBHOOK: null,
-    RECEIVED_COMMANDS: null,
-    RECEIVED_COMMAND: null,
-    UPDATED_COMMAND: null,
-    REMOVED_COMMAND: null,
-    RECEIVED_OAUTHAPPS: null,
-    RECEIVED_OAUTHAPP: null,
-    REMOVED_OAUTHAPP: null,
-
-    RECEIVED_CUSTOM_EMOJIS: null,
-    RECEIVED_CUSTOM_EMOJI: null,
-    UPDATED_CUSTOM_EMOJI: null,
-    REMOVED_CUSTOM_EMOJI: null,
-
-    RECEIVED_REACTIONS: null,
-    ADDED_REACTION: null,
-    REMOVED_REACTION: null,
-
-    RECEIVED_MSG: null,
-
-    RECEIVED_TEAM: null,
-    RECEIVED_MY_TEAM: null,
-    CREATED_TEAM: null,
-    UPDATE_TEAM: null,
-
     SET_NAVIGATION_BLOCKED: null,
     DEFER_NAVIGATION: null,
     CANCEL_NAVIGATION: null,
     CONFIRM_NAVIGATION: null,
-    RECEIVED_CONFIG: null,
-    RECEIVED_LOGS: null,
-    RECEIVED_SERVER_AUDITS: null,
-    RECEIVED_SERVER_COMPLIANCE_REPORTS: null,
-    RECEIVED_ALL_TEAMS: null,
-    RECEIVED_ALL_TEAM_LISTINGS: null,
-    RECEIVED_MY_TEAM_MEMBERS: null,
-    RECEIVED_MY_TEAMS_UNREAD: null,
-    RECEIVED_MEMBERS_IN_TEAM: null,
-    RECEIVED_TEAM_STATS: null,
-
-    UPDATE_OPEN_GRAPH_METADATA: null,
-    RECIVED_OPEN_GRAPH_METADATA: null,
-
-    SHOW_SEARCH: null,
 
     TOGGLE_SHORTCUTS_MODAL: null,
     TOGGLE_IMPORT_THEME_MODAL: null,
@@ -224,16 +133,7 @@ export const ActionTypes = keyMirror({
     SHOW_EDIT_POST_MODAL: null,
     HIDE_EDIT_POST_MODAL: null,
 
-    SUGGESTION_PRETEXT_CHANGED: null,
-    SUGGESTION_RECEIVED_SUGGESTIONS: null,
-    SUGGESTION_CLEAR_SUGGESTIONS: null,
-    SUGGESTION_COMPLETE_WORD: null,
-    SUGGESTION_SELECT_NEXT: null,
-    SUGGESTION_SELECT_PREVIOUS: null,
-
     BROWSER_CHANGE_FOCUS: null,
-
-    EMOJI_POSTED: null,
 
     RECEIVED_PLUGIN_COMPONENT: null,
     REMOVED_PLUGIN_COMPONENT: null,
@@ -245,8 +145,6 @@ export const ActionTypes = keyMirror({
 
     MODAL_OPEN: null,
     MODAL_CLOSE: null,
-
-    POPOVER_MENTION_KEY_CLICK: null,
 
     SELECT_CHANNEL_WITH_MEMBER: null,
